Change the epsilon to a sequential solve
Solve two sequential problem, the first optimise the actual objective function and the second maximise all stats. This ensure that the solver does not lie, and the solution is more elegant than adding an epsilon to all weights (since zero weights made the solver lie when it had nothing to gain by presenting the true value).